+++ changelog.md
@@ -5,6 +5,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 # [Unreleased] - 2024-05-27
+* Fixed - Properly read changes and restart stack after using the `slic php-version set` command.
 * Added - PHP 8.3 docker images.
 * Change - Update GitHub action versions to remove node notices.
 * Change - Build slic images from the latest PHP patch versions.